fix cyclic dependencies
