+++ libs/backendapi/backend.go
@@ -1,7 +1,7 @@
 package backendapi
 
 import (
-	"github.com/diggerhq/digger/cli/pkg/core/execution"
+	"github.com/diggerhq/digger/libs/execution"
 	"github.com/diggerhq/digger/libs/scheduler"
 	"time"
 )
